Renamed service namespace to api
